AMP-695: - refactor dropbox related methods (move from fileStorageService to DropboxService) - add code for creating dropbox dirs per colelction and for all collection
