+++ src/main/java/edu/indiana/dlib/amppd/service/impl/FileStorageServiceImpl.java
@@ -399,32 +399,4 @@ public class FileStorageServiceImpl implements FileStorageService {
 		}
 	}
 	
-	public String encodeUri(String path) {
-		Pattern pattern = Pattern.compile("[^A-Za-z0-9._-]");
-        
-		char[] chars = path.toCharArray();
-
-		for (char ch : chars) {
-			String charString = Character.toString(ch);
-	        Matcher matcher = pattern.matcher(charString);
-	        while (matcher.find()) {
-	            String hexString = "%" + Integer.toHexString((int) ch);
-	            path = path.replace(charString, hexString);
-	        }
-		}		
-		
-	    return path;
-	}
-	
-	public Path getDropboxPath(String unitName, String collectionName) {
-		Path unitPath = getDropboxPath(unitName);
-		String encodedCollectionName = encodeUri(collectionName);
-		System.out.println("Collection Name: "  + encodedCollectionName);
-		return Paths.get(unitPath.toString(), encodedCollectionName);
-	}
-	
-	public Path getDropboxPath(String unitName) {
-		String encodedUnitName = encodeUri(unitName);
-		return Paths.get(config.getDropboxRoot(), encodedUnitName);
-	}
 }
